+++ ms2/src/org/labkey/ms2/MS2Controller.java
@@ -362,7 +362,7 @@ public class MS2Controller extends SpringActionController
     }
 
     /** @return URL with .lastFilter if user has configured their default view that way and with a run id */
-    public static ActionURL getShowRunURL(User user, Container c, int runId)
+    public static ActionURL getShowRunURL(User user, Container c, long runId)
     {
         ActionURL url = getShowRunURL(user, c);
         url.addParameter(RunForm.PARAMS.run, String.valueOf(runId));
@@ -2062,9 +2062,9 @@ public class MS2Controller extends SpringActionController
         public boolean handlePost(FormType form, BindException errors) throws RunListException
         {
             ActionURL currentURL = getViewContext().getActionURL();
-            int runListId = RunListCache.cacheSelectedRuns(getRequiresSameType(), form, getViewContext());
+            long runListId = RunListCache.cacheSelectedRuns(getRequiresSameType(), form, getViewContext());
             _successUrl = currentURL.clone();
-            _successUrl.addParameter("runList", Integer.toString(runListId));
+            _successUrl.addParameter("runList", Long.toString(runListId));
 
             return true;
         }
@@ -3322,7 +3322,7 @@ public class MS2Controller extends SpringActionController
         @Override
         public ModelAndView getView(DetailsForm form, BindException errors)
         {
-            int runId;
+            long runId;
             int seqId;
             if (form.run != 0)
             {
@@ -3909,7 +3909,7 @@ public class MS2Controller extends SpringActionController
     public static class SeqRunIdPair
     {
         private int _seqId;
-        private int _run;
+        private long _run;
 
         public void setSeqId(int seqId)
         {
@@ -3921,12 +3921,12 @@ public class MS2Controller extends SpringActionController
             return _seqId;
         }
 
-        public void setRun(int run)
+        public void setRun(long run)
         {
             _run = run;
         }
 
-        public int getRun()
+        public long getRun()
         {
             return _run;
         }
@@ -4585,7 +4585,7 @@ public class MS2Controller extends SpringActionController
             run, expanded, grouping, highestScore
         }
 
-        int run = 0;
+        long run = 0;
         int fraction = 0;
         int tryptic;
         boolean expanded = false;
@@ -4616,12 +4616,12 @@ public class MS2Controller extends SpringActionController
             return this.highestScore;
         }
 
-        public void setRun(int run)
+        public void setRun(long run)
         {
             this.run = run;
         }
 
-        public int getRun()
+        public long getRun()
         {
             return run;
         }
