+++ ms2/src/org/labkey/ms2/PepXmlExperimentDataHandler.java
@@ -181,7 +181,7 @@ public class PepXmlExperimentDataHandler extends AbstractExperimentDataHandler
     }
 
     @Override
-    public void runMoved(ExpData newData, Container container, Container targetContainer, String oldRunLSID, String newRunLSID, User user, int oldDataRowID) throws ExperimentException
+    public void runMoved(ExpData newData, Container container, Container targetContainer, String oldRunLSID, String newRunLSID, User user, long oldDataRowID) throws ExperimentException
     {
         super.runMoved(newData, container, targetContainer, oldRunLSID, newRunLSID, user, oldDataRowID);
         updateRunLSID(newData, container, targetContainer, newRunLSID, user);
